backend/requirements.txt
# FastAPI 웹 프레임워크
fastapi==0.115.0

# ASGI 서버 (FastAPI 실행용)
uvicorn[standard]==0.30.6

# DB ORM
SQLAlchemy==2.0.34

# PostgreSQL 드라이버
psycopg2-binary==2.9.9

# 데이터 처리
pandas==2.2.3
openpyxl==3.1.5   # Excel 읽기
xlrd==2.0.1       # (구버전 Excel 호환 필요 시)

# Pydantic (데이터 검증, FastAPI 스키마)
pydantic==2.9.2
pydantic-settings==2.4.0   # .env 관리용

# 테스트용
pytest==8.3.3
httpx==0.27.2
